+++ include/f_op/f_op_actor.h
@@ -329,9 +329,9 @@ public:
     cXyz& getHeadLockPos() { return mHeadLockPos; }
 
     void onCutDownHitFlg() { mFlags |= (u16)fopEn_flag_CutDownHit; }
-    void onWolfBiteDamage() { mFlags |= fopEn_flag_WolfBiteDamage; }
-    void onWolfDownStartFlg() { mFlags |= (fopEn_flag_WolfDownPull | fopEn_flag_WolfDownStart); }
-    void onWolfDownPullEndFlg() { mFlags |= fopEn_flag_WolfDownPullEnd; }
+    void onWolfBiteDamage() { mFlags |= (u16)fopEn_flag_WolfBiteDamage; }
+    void onWolfDownStartFlg() { mFlags |= (u16)(fopEn_flag_WolfDownPull | fopEn_flag_WolfDownStart); }
+    void onWolfDownPullEndFlg() { mFlags |= (u16)fopEn_flag_WolfDownPullEnd; }
     void onWolfNoLock() { mFlags |= (u16)fopEn_flag_WolfNoLock; }
     void onDownFlg() { mFlags |= (u16)fopEn_flag_Down; }
     void onHeadLockFlg() { mFlags |= (u16)fopEn_flag_HeadLock; }
